# docs/openapi.yaml
openapi: 3.1.0
info:
  title: S-SPARC AI & E-STRANGE Platform - API Reference
  version: 3.0.0
  summary: Sustainable Smart Personal Assistant for Responsible Consumption & Code Evaluation Microservice
  description: "# S-SPARC AI & E-STRANGE Platform Documentation\n\nWelcome to the official developer documentation and API reference for **S-SPARC (Sustainable Smart Personal Assistant for Responsible Consumption)** integrated with the **E-STRANGE Learning Management Platform**.\n\n---\n\n## System Architecture Overview\n\nS-SPARC is engineered around 4 core operational tiers designed for high-concurrency academic software engineering tutoring:\n\n1. **E-STRANGE Parent LMS (PHP Platform)**:\n   - Handles student course enrollment, assignment submissions, automated similarity analysis (`suspicion`), peer reviews (`code_clarity_suggestion`), gamification leaderboards, and grade tracking.\n2. **S-SPARC AI Frontend (PHP/JS Client)**:\n   - Interactive eco-aware coding workspace featuring token usage gauges, real-time AI responses, and environmental footprint visualizers.\n3. **S-SPARC FastAPI Backend (Python Core)**:\n   - Core API layer delivering user session auth, semantic search cache (0\
...
  contact:
    name: S-SPARC Engineering & Research Team
    email: support@s-sparc.ac.id
    url: http://localhost:5000/redocly
  license:
    name: Academic Research & Sustainability License 1.0
    url: https://opensource.org/licenses/MIT
servers:
- url: http://127.0.0.1:5000
...
paths:
...
  /register:
    post:
      tags:
      - Authentication
      summary: Register New Account
      description: Registers a new student or instructor user account with SHA-256 password hashing.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: mahasiswa_sipil_2026
              email: mahasiswa@s-sparc.ac.id
              password: RahasiaSuper123!
      responses:
        '201':
          description: User account registered successfully.
          content:
            application/json:
              example:
                message: Registrasi berhasil.
        '400':
          description: Missing required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username or email already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      tags:
      - Authentication
      summary: User Login & Session Establishment
      description: Authenticates credentials against database records, establishing a secure HTTP-only session cookie.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: mahasiswa_sipil_2026
              password: RahasiaSuper123!
      responses:
        '200':
          description: Login successful. Session cookie issued.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HTTP-only session cookie matching Flask/PHP session key.
          content:
            application/json:
              example:
                message: Login berhasil.
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /whoami:
    get:
      tags:
      - Authentication
      summary: Get Logged-In User Profile
      description: Retrieves user details of the active authenticated session.
      operationId: whoAmI
      security:
      - SessionCookieAuth: []
      - HeaderUserIdAuth: []
      responses:
        '200':
          description: Profile data retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              example:
                user_id: usr-8f3a-9901-b2c3
                username: mahasiswa_sipil_2026
                email: mahasiswa@s-sparc.ac.id
                is_admin: 0
        '401':
          description: Unauthorized access.
  /change-password:
    post:
      tags:
      - Authentication
      summary: Change User Password
      description: Allows logged-in users to update password by verifying current credentials.
      operationId: changePassword
      security:
      - SessionCookieAuth: []
      - HeaderUserIdAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password updated successfully.
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: HTTP-only session cookie established via POST /login.
    HeaderUserIdAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Explicit user ID header.
  schemas:
...
    RegisterRequest:
      type: object
      required:
      - username
      - email
      - password
      properties:
        username:
          type: string
          example: mahasiswa_sipil_2026
        email:
          type: string
          example: mahasiswa@s-sparc.ac.id
        password:
          type: string
          example: RahasiaSuper123!
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          example: mahasiswa_sipil_2026
        password:
          type: string
          example: RahasiaSuper123!
    ChangePasswordRequest:
      type: object
      required:
      - old_password
      - new_password
      properties:
        old_password:
          type: string
        new_password:
          type: string
...
    UserProfileResponse:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        email:
          type: string
        is_admin:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: Terjadi kesalahan pada server.